调用sidecar接口权限，限制为管理员
https://github.com/steedos/steedos-platform/issues/2845

+++ services/service-sidercar/package.service.js
@@ -51,7 +51,12 @@ module.exports = {
 			params: {
 			},
 			async handler(ctx) {
-				return await ctx.call('$node.list');
+				try {
+					await this.checkIsAdmin(ctx);
+					return await ctx.call('$node.list');
+				} catch (error) {
+					return this.onError(ctx, error);
+				}
 			}
 		},
 		/**
@@ -65,7 +70,12 @@ module.exports = {
 			params: {
 			},
 			async handler(ctx) {
-				return await ctx.call('$node.services');
+				try {
+					await this.checkIsAdmin(ctx);
+					return await ctx.call('$node.services');
+				} catch (error) {
+					return this.onError(ctx, error);
+				}
 			}
 		},
 		/**
@@ -79,7 +89,12 @@ module.exports = {
 			params: {
 			},
 			async handler(ctx) {
-				return await ctx.call('$node.actions');
+				try {
+					await this.checkIsAdmin(ctx);
+					return await ctx.call('$node.actions');
+				} catch (error) {
+					return this.onError(ctx, error);
+				}
 			}
 		},
 		/**
@@ -93,7 +108,12 @@ module.exports = {
 			params: {
 			},
 			async handler(ctx) {
-				return await ctx.call('$node.events');
+				try {
+					await this.checkIsAdmin(ctx);
+					return await ctx.call('$node.events');
+				} catch (error) {
+					return this.onError(ctx, error);
+				}
 			}
 		},
 
@@ -109,67 +129,72 @@ module.exports = {
 				name: "string|no-empty"
 			},
 			async handler(ctx) {
-				const schema = _.cloneDeep(ctx.params);
+				try {
+					await this.checkIsAdmin(ctx);
+					const schema = _.cloneDeep(ctx.params);
 
-				// 1. Check the service is exist
-				let svc = this.broker.getLocalService({
-					name: schema.name,
-					version: schema.version
-				});
-
-				// 2.   If yes, destroy
-				if (svc) {
-					this.logger.info(`Destroy previous '${schema.name}' service...`);
-					await this.broker.destroyService(svc);
-				}
-
-				// 3. Convert the schema, fulfill the action/event handlers
-				this.logger.info(kleur.yellow().bold(`Create new '${schema.name}' service...`));
-
-				if (schema.actions) {
-					Object.entries(schema.actions).forEach(([name, action]) => {
-						// Handle shorthand action def
-						if (typeof action == "string") {
-							schema.actions[name] = { handler: action };
-							action = schema.actions[name];
-						}
-
-						if (typeof action.handler != "string") {
-							throw new MoleculerClientError(
-								"Invalid action handler",
-								400,
-								"INVALID_HANDLER",
-								{ action }
-							);
-						}
-
-						action.handler = this.generateActionHandler(action, schema);
+					// 1. Check the service is exist
+					let svc = this.broker.getLocalService({
+						name: schema.name,
+						version: schema.version
 					});
+
+					// 2.   If yes, destroy
+					if (svc) {
+						this.logger.info(`Destroy previous '${schema.name}' service...`);
+						await this.broker.destroyService(svc);
+					}
+
+					// 3. Convert the schema, fulfill the action/event handlers
+					this.logger.info(kleur.yellow().bold(`Create new '${schema.name}' service...`));
+
+					if (schema.actions) {
+						Object.entries(schema.actions).forEach(([name, action]) => {
+							// Handle shorthand action def
+							if (typeof action == "string") {
+								schema.actions[name] = { handler: action };
+								action = schema.actions[name];
+							}
+
+							if (typeof action.handler != "string") {
+								throw new MoleculerClientError(
+									"Invalid action handler",
+									400,
+									"INVALID_HANDLER",
+									{ action }
+								);
+							}
+
+							action.handler = this.generateActionHandler(action, schema);
+						});
+					}
+
+					if (schema.events) {
+						Object.entries(schema.events).forEach(([name, event]) => {
+							// Handle shorthand event def
+							if (typeof event == "string") {
+								schema.events[name] = { handler: event };
+								event = schema.events[name];
+							}
+							if (typeof event.handler != "string") {
+								throw new MoleculerClientError(
+									"Invalid event handler",
+									400,
+									"INVALID_HANDLER",
+									{ event }
+								);
+							}
+
+							event.handler = this.generateEventHandler(event, schema);
+						});
+					}
+
+					svc = this.broker.createService(schema);
+
+					return { status: "OK" };
+				} catch (error) {
+					return this.onError(ctx, error);
 				}
-
-				if (schema.events) {
-					Object.entries(schema.events).forEach(([name, event]) => {
-						// Handle shorthand event def
-						if (typeof event == "string") {
-							schema.events[name] = { handler: event };
-							event = schema.events[name];
-						}
-						if (typeof event.handler != "string") {
-							throw new MoleculerClientError(
-								"Invalid event handler",
-								400,
-								"INVALID_HANDLER",
-								{ event }
-							);
-						}
-
-						event.handler = this.generateEventHandler(event, schema);
-					});
-				}
-
-				svc = this.broker.createService(schema);
-
-				return { status: "OK" };
 			}
 		},
 
@@ -185,23 +210,27 @@ module.exports = {
 				serviceName: "string|no-empty|trim"
 			},
 			async handler(ctx) {
-				// 1. Check the service is exist
-				const svc = this.broker.getLocalService(ctx.params.serviceName);
+				try {
+					await this.checkIsAdmin(ctx);
+					// 1. Check the service is exist
+					const svc = this.broker.getLocalService(ctx.params.serviceName);
 
-				// 2.   If yes, destroy
-				if (!svc) {
-					throw new MoleculerClientError(
-						"Service not found",
-						404,
-						"NOT_FOUND",
-						ctx.params
-					);
+					// 2.   If yes, destroy
+					if (!svc) {
+						throw new MoleculerClientError(
+							"Service not found",
+							404,
+							"NOT_FOUND",
+							ctx.params
+						);
+					}
+					this.logger.info(kleur.yellow().bold(`Destroy '${svc.fullName}' service...`));
+					await this.broker.destroyService(svc);
+
+					return { status: "OK" };
+				} catch (error) {
+					return this.onError(ctx, error);
 				}
-
-				this.logger.info(kleur.yellow().bold(`Destroy '${svc.fullName}' service...`));
-				await this.broker.destroyService(svc);
-
-				return { status: "OK" };
 			}
 		},
 
@@ -220,38 +249,44 @@ module.exports = {
 				options: "object|optional|default:{}"
 			},
 			async handler(ctx) {
-				const payload = ctx.params;
 				try {
-					const response = await ctx.call(
-						payload.action,
-						payload.params != null ? payload.params : {},
-						{
-							meta: payload.meta,
-							timeout: payload.options.timeout,
-							retries: payload.options.retries
-						}
-					);
+					await this.checkIsAdmin(ctx);
+					const payload = ctx.params;
+					try {
+						const response = await ctx.call(
+							payload.action,
+							payload.params != null ? payload.params : {},
+							{
+								meta: payload.meta,
+								timeout: payload.options.timeout,
+								retries: payload.options.retries
+							}
+						);
 
-					return {
-						response,
-						meta: ctx.meta
-					};
-				} catch (err) {
-					ctx.meta.$statusCode = err.code || 500;
+						return {
+							response,
+							meta: ctx.meta
+						};
+					} catch (err) {
+						ctx.meta.$statusCode = err.code || 500;
 
-					return {
-						error: _.pick(err, [
-							"name",
-							"message",
-							"code",
-							"type",
-							"stack",
-							"data",
-							"nodeID"
-						]),
-						meta: ctx.meta
-					};
+						return {
+							error: _.pick(err, [
+								"name",
+								"message",
+								"code",
+								"type",
+								"stack",
+								"data",
+								"nodeID"
+							]),
+							meta: ctx.meta
+						};
+					}
+				} catch (error) {
+					return this.onError(ctx, error);
 				}
+
 			}
 		},
 
@@ -270,14 +305,19 @@ module.exports = {
 				options: "object|optional|default:{}"
 			},
 			async handler(ctx) {
-				const payload = ctx.params;
 				try {
-					await ctx.emit(payload.event, payload.params != null ? payload.params : {}, {
-						meta: payload.meta,
-						groups: payload.options.groups
-					});
-				} catch (err) {
-					this.logger.error("Unable to emit event", err);
+					await this.checkIsAdmin(ctx);
+					const payload = ctx.params;
+					try {
+						await ctx.emit(payload.event, payload.params != null ? payload.params : {}, {
+							meta: payload.meta,
+							groups: payload.options.groups
+						});
+					} catch (err) {
+						this.logger.error("Unable to emit event", err);
+					}
+				} catch (error) {
+					return this.onError(ctx, error);
 				}
 			}
 		},
@@ -297,14 +337,19 @@ module.exports = {
 				options: "object|optional|default:{}"
 			},
 			async handler(ctx) {
-				const payload = ctx.params;
 				try {
-					await ctx.broadcast(payload.event, payload.params, {
-						meta: payload.meta,
-						groups: payload.options.groups
-					});
-				} catch (err) {
-					this.logger.error("Unable to broadcast event", err);
+					await this.checkIsAdmin(ctx);
+					const payload = ctx.params;
+					try {
+						await ctx.broadcast(payload.event, payload.params, {
+							meta: payload.meta,
+							groups: payload.options.groups
+						});
+					} catch (err) {
+						this.logger.error("Unable to broadcast event", err);
+					}
+				} catch (error) {
+					return this.onError(ctx, error);
 				}
 			}
 		}
@@ -404,6 +449,21 @@ module.exports = {
 
 				return payload;
 			});
+		},
+
+		async checkIsAdmin(ctx) {
+			const user = ctx.meta.user;
+			if (!user || !user.is_space_admin) {
+				let err = new Error('NO_RIGHTS');
+				err.code = 401;
+				throw err;
+			}
+		},
+
+		onError(ctx, error) {
+			const msg = error.message;
+			ctx.meta.$statusCode = error.code || 500;
+			return msg
 		}
 	},
